fix: useless conditional (#1631)
* fix: useless conditional

Error spotted by VS Code Advanced Security:

If a condition always evaluates to true or always evaluates to false, this often indicates incomplete code or a latent bug and should be examined carefully.

Recommendation
Examine the surrounding code to determine why the condition is useless. If it is no longer needed, remove it.

* Update htmlhint.js